backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-auth-oauthlib==1.0.0
google-auth==2.26.0
razorpay==1.3.0
gunicorn==21.2.0
